api/requirements.txt
# Django
Django==6.0.2
djangorestframework==3.16.0
django-import-export==4.3.6
django-qsessions==2.1.0
python-decouple==3.8
django-redis
celery[redis]>=5.6
flower>=2.0
django-mongodb-backend==6.0.2
django-cors-headers==4.9.0

# Deployment
uvicorn==0.34.0
whitenoise==6.11.0
sentry-sdk
# Ingestion
requests>=2.32.5
feedparser>=6.0
telethon>=1.36
transformers>=5.3
sentence-transformers>=3.0
vaderSentiment>=3.3
openai>=2.24.0

# HTML parsing
scrapling>=0.2.9

# Geocoding
geonamescache>=3.0.0

# Email
boto3>=1.37.0
markdown>=3.7

# Stream data fetching (prices, forex, NOTAM, earthquakes)
yfinance>=1.2.0

# PyTorch CPU-only (avoids pulling CUDA binaries)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.10.0+cpu
